+++ docs/changelog.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/), 
 Version numbering adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - [master] - 2025-??-??
+  - [Added]   *Human match* versus *Exact match* concept added, with `-em`, `-ic`, `-iw` and `-ibl` options
   - [Added]   keyword `executable` added to create scripts
   - [Added]   "Crate of the year" and tests results badges added
   - [changed] Close #5 and #15 (more clear msg on spawn problems and auto find the exe in PATH)